requirements.txt
# requirements.txt
snowflake-connector-python
streamlit_autorefresh
streamlit_apexjs